Move script/fetch-docs.js -> lib/fetch-docs.js
