+++ lib/fetch-docs.js
